fixup! cinnamon.gnome-online-accounts-gtk: init at 0-unstable-2024-03-23
This is an oversight.

When I saw the "Sign in to %s with your browser" dialog I really thought
this is not needed, but no, there are (still) usages of g_network_monitor_get_default
here, taking you to the GDummyTlsBackend when glib-networking is missing,
which is the reason we see "TLS support is not available" in c-c-c.

https://github.com/NixOS/nixpkgs/commit/47abf03340330b4748b68abc846067713e2e4ed6

Reported-by: aleksana <me@aleksana.moe>